+++ front/src/app/Community/PostCreate/page.tsx
@@ -1,6 +1,5 @@
 import Nav from "@/app/components/Nav";
 import PostCreates from "@/app/components/PostCreates";
-import PictureUpload from "@/app/components/PictureUpload";
 
 export default function PostCreate() {
   return (
@@ -8,7 +7,6 @@ export default function PostCreate() {
       <Nav />
       <div className="flex justify-center w-full h-auto">
         <div className="flex-col items-center min-w-[700px] max-w-[1000px] w-11/12 h-auto">
-          <PictureUpload />
           <PostCreates />
         </div>
       </div>
